+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 - Abort directly if a command fails in do review
 - Execute gg do commit after localizing in kidney add
 - Pass gitRef param to \_localizeRefs.get in DoReviewCommand and tests
+- open ticket as VSCode workspace file instead of individual repos
 
 ### Removed